requirements.txt
langchain
langchain-openai
langchain-mcp-adapters
PyJWT[crypto]
requests